requirements.txt
Django==5.2.7
djangorestframework==3.16.1
psycopg2-binary==2.9.11
httpx==0.28.1
pytest-django==4.11.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
python-dotenv==1.1.1
